requirements.txt
pyTelegramBotAPI==4.14.1
pytz==2023.3
Levenshtein==0.20.8
matplotlib==3.6.3
urllib3==1.26.12
aiohttp==3.8.5
python-dotenv==1.0.0


